agent/pyproject.toml
[tool.poetry]
name = "repository_agent"
version = "0.1.0"
description = "Repository Analysis Agent"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[project]
name = "repository_agent"
version = "0.0.1"
dependencies = [
  "langgraph",
  "langchain_core",
  "langchain_openai",
  "langchain-google-genai",
  "langchain",
  "openai",
  "langchain-community",
  "tavily-python",
  "python-dotenv",
  "uvicorn",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.39"
langchain-core = "^0.3.12"
langchain-openai = "0.2.3"
langchain-anthropic = "0.2.3"
langchain-google-genai = "2.0.0"
langchain = "0.3.4"
openai = "^1.52.1"
langchain-community = "^0.3.1"
tavily-python = "^0.5.0"
copilotkit = "0.1.28"
python-dotenv = "^1.0.1"
uvicorn = "^0.31.0"

[tool.poetry.scripts]
start = "repository_agent.__main__:main" 